% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_kivt.R
\name{compute_kivt}
\alias{compute_kivt}
\title{Compute Kivt coefficient}
\usage{
compute_kivt(calendar, clusters_desc)
}
\arguments{
\item{calendar}{Calendar data read with \code{\link{read_calendar}}.}

\item{clusters_desc}{Clusters / groups description read with \code{\link{read_cluster_desc}}.}
}
\value{
a \code{data.table}
}
\description{
Compute Kivt coefficient
}
\examples{
\dontrun{
# Calendar data
calendar <- read_calendar(path = "REF_Planning_5_ans_mars_2018.xlsx")

# Clusters description
clusters <- read_cluster_desc("HypothesesRTE_CHO-4145.xlsx")


# Kivt computation
kivt <- compute_kivt(calendar, clusters)
kivt
}
}
