% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clusters.R
\name{create_clusters}
\alias{create_clusters}
\title{Create clusters}
\usage{
create_clusters(calendar, opts = simOptions())
}
\arguments{
\item{calendar}{Calendar data read with \code{read_calendar}.}

\item{opts}{List of simulation parameters returned by the function
\code{setSimulationPath}}
}
\description{
Create clusters
}
