% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-kd-cho.R
\name{read_kd_cho}
\alias{read_kd_cho}
\title{Read sheet "Kd CHO" from RTE hypothesis}
\usage{
read_kd_cho(path)
}
\arguments{
\item{path}{Path to Excel file.}
}
\value{
a \code{data.table}
}
\description{
Read sheet "Kd CHO" from RTE hypothesis
}
